Remove Unnecessary Abstraction
