+++ libs/platform/src/banner/banner.service.ts
@@ -1,6 +1,24 @@
-import { Observable } from "rxjs";
+import { Observable, map } from "rxjs";
 
-export abstract class BannerService {
+import {
+  ActiveUserState,
+  BANNER_DISK,
+  KeyDefinition,
+  StateProvider,
+} from "@bitwarden/common/platform/state";
+
+const STATES_KEY = KeyDefinition.record<boolean>(BANNER_DISK, "states", {
+  deserializer: (b) => b,
+});
+
+export class BannerService {
+  private bannerStates: ActiveUserState<Record<string, boolean>>;
+
+  constructor(private stateProvider: StateProvider) {
+    this.bannerStates = this.stateProvider.getActive(STATES_KEY);
+    this.bannerStates$ = this.bannerStates.state$;
+  }
+
   /**
    * An observable containing all banners in a record with their id and the visibility.
    */
@@ -10,12 +28,19 @@ export abstract class BannerService {
    * @param bannerId The banner id of the visibility to track
    * @returns An observable boolean or undefined showing that state of the observable
    */
-  bannerVisibility$: (bannerId: string) => Observable<boolean | undefined>;
+  bannerVisibility$(bannerId: string) {
+    return this.bannerStates.state$.pipe(map((states) => states[bannerId]));
+  }
 
   /**
    * A method to set the visibility of a banner via its id.
    * @param bannerId The banner id of the visibility to update.
    * @param visibility A boolean where true means the banner should be visible and false is the banner should be not be visible.
    */
-  setBannerVisibility: (bannerId: string, visibility: boolean) => Promise<void>;
+  async setBannerVisibility(bannerId: string, visibility: boolean) {
+    await this.bannerStates.update((states) => {
+      states[bannerId] = visibility;
+      return states;
+    });
+  }
 }
