+++ libs/platform/src/index.ts
@@ -1,3 +1,1 @@
 export * from "./banner/banner.service";
-// TODO: Make this harder to import, it's only imported to be used in DI.
-export * from "./banner/implementation/default-banner.service";
